+++ packages/atomic/cypress/integration/facets/timeframe-facet/timeframe-facet.cypress.ts
@@ -41,10 +41,6 @@ describe('Timeframe Facet V1 Test Suites', () => {
         TimeframeFacetSelectors,
         0
       );
-      CommonFacetAssertions.assertNumberOfIdleLinkValues(
-        TimeframeFacetSelectors,
-        defaultNumberOfValues
-      );
       CommonFacetAssertions.assertDisplayClearButton(
         TimeframeFacetSelectors,
         false
@@ -62,8 +58,7 @@ describe('Timeframe Facet V1 Test Suites', () => {
       describe('verify rendering', () => {
         before(setupSelectLinkValue);
         CommonFacetAssertions.assertDisplayFacet(TimeframeFacetSelectors, true);
-        // CommonAssertions.assertAccessibility(TimeframeFacetSelectors);
-        // TODO: enable it when KIT-996 fixed
+        CommonAssertions.assertAccessibility(timeframeFacetComponent);
         CommonFacetAssertions.assertDisplayClearButton(
           TimeframeFacetSelectors,
           true
@@ -97,7 +92,7 @@ describe('Timeframe Facet V1 Test Suites', () => {
 
         describe('verify rendering', () => {
           before(setupSecondLinkValue);
-          // CommonAssertions.assertAccessibility(timeframeFacetComponent);
+          CommonAssertions.assertAccessibility(timeframeFacetComponent);
           CommonFacetAssertions.assertNumberOfSelectedLinkValues(
             TimeframeFacetSelectors,
             1
@@ -108,7 +103,7 @@ describe('Timeframe Facet V1 Test Suites', () => {
           );
         });
 
-        describe('versify analytics', () => {
+        describe('verify analytics', () => {
           before(setupSecondLinkValue);
           TimeframeFacetAssertions.assertLogTimeframeFacetSelect(
             field,
@@ -198,7 +193,6 @@ describe('Timeframe Facet V1 Test Suites', () => {
 
   describe('with custom #field timeframe-facet', () => {
     const customField = 'indexeddate';
-    const defaultNumberOfValues = unitFrames.length;
     function setupTimeframeFacet() {
       new TestFixture()
         .with(addTimeframeFacet({label, field: customField}, unitFrames))
@@ -214,41 +208,6 @@ describe('Timeframe Facet V1 Test Suites', () => {
         false
       );
       CommonFacetAssertions.assertLabelContains(TimeframeFacetSelectors, label);
-    });
-
-    describe('when selecting a value', () => {
-      const selectionIndex = 1;
-      function setupSelectLinkValue() {
-        setupTimeframeFacet();
-        selectIdleLinkValueAt(TimeframeFacetSelectors, selectionIndex);
-        cy.wait(TestFixture.interceptAliases.Search);
-      }
-
-      describe('verify rendering', () => {
-        before(setupSelectLinkValue);
-        CommonFacetAssertions.assertDisplayFacet(TimeframeFacetSelectors, true);
-        // CommonAssertions.assertAccessibility(TimeframeFacetSelectors);
-        CommonFacetAssertions.assertDisplayClearButton(
-          TimeframeFacetSelectors,
-          true
-        );
-        CommonFacetAssertions.assertNumberOfSelectedLinkValues(
-          TimeframeFacetSelectors,
-          1
-        );
-        CommonFacetAssertions.assertNumberOfIdleLinkValues(
-          TimeframeFacetSelectors,
-          defaultNumberOfValues - 1
-        );
-      });
-
-      describe('verify analytic', () => {
-        before(setupSelectLinkValue);
-        TimeframeFacetAssertions.assertLogTimeframeFacetSelect(
-          customField,
-          unitFrames[selectionIndex].unit
-        );
-      });
     });
   });
 
@@ -319,14 +278,13 @@ describe('Timeframe Facet V1 Test Suites', () => {
   });
 
   describe('with custom #amount timeframe', () => {
-    const amount = 10;
     const periodFrames = [
-      {unit: 'day', amount: amount},
-      {unit: 'week', amount: amount},
-      {unit: 'month', amount: amount},
+      {unit: 'month', amount: 20},
+      {unit: 'year', amount: 3},
+      {unit: 'year', amount: 5},
     ];
     const customNumberOfValues = periodFrames.length;
-    const selectionIndex = 1;
+    const selectionIndex = 0;
     function setupTimeframeFacetCustomAmount() {
       new TestFixture()
         .with(addTimeframeFacet({label, field}, periodFrames))
@@ -348,57 +306,14 @@ describe('Timeframe Facet V1 Test Suites', () => {
       );
       TimeframeFacetAssertions.assertFacetValueContainsText(
         selectionIndex,
-        `Past ${amount} ${periodFrames[selectionIndex].unit}s`
+        `Past ${periodFrames[selectionIndex].amount} ${periodFrames[selectionIndex].unit}s`
       );
-    });
-
-    describe('when selecting a value', () => {
-      function setupSelectLinkValue() {
-        setupTimeframeFacetCustomAmount();
-        selectIdleLinkValueAt(TimeframeFacetSelectors, selectionIndex);
-        cy.wait(TestFixture.interceptAliases.Search);
-      }
-
-      describe('verify rendering', () => {
-        before(setupSelectLinkValue);
-        CommonFacetAssertions.assertDisplayFacet(TimeframeFacetSelectors, true);
-        // CommonAssertions.assertAccessibility(TimeframeFacetSelectors);
-        CommonFacetAssertions.assertDisplayClearButton(
-          TimeframeFacetSelectors,
-          true
-        );
-        CommonFacetAssertions.assertNumberOfSelectedLinkValues(
-          TimeframeFacetSelectors,
-          1
-        );
-        CommonFacetAssertions.assertNumberOfIdleLinkValues(
-          TimeframeFacetSelectors,
-          customNumberOfValues - 1
-        );
-      });
-
-      describe('verify analytic', () => {
-        before(setupSelectLinkValue);
-        TimeframeFacetAssertions.assertLogTimeframeFacetSelect(
-          field,
-          periodFrames[selectionIndex].unit,
-          'past',
-          periodFrames[selectionIndex].amount
-        );
-      });
     });
   });
 
-  describe.skip('with custom #period timeframe', () => {
-    // TODO: enable it when KIT-995 fixed
+  describe('with custom #period timeframe', () => {
     const period = 'next';
-    const periodFrames = [
-      {unit: 'day', period: period},
-      {unit: 'week', period: period},
-      {unit: 'month', period: period},
-    ];
-    const customNumberOfValues = periodFrames.length;
-    const selectionIndex = 1;
+    const periodFrames = [{unit: 'year', period, amount: 25}];
     function setupTimeframeFacetCustomAmount() {
       new TestFixture()
         .with(addTimeframeFacet({label, field}, periodFrames))
@@ -409,53 +324,27 @@ describe('Timeframe Facet V1 Test Suites', () => {
       before(setupTimeframeFacetCustomAmount);
       CommonFacetAssertions.assertDisplayFacet(TimeframeFacetSelectors, true);
       CommonAssertions.assertAccessibility(timeframeFacetComponent);
-      CommonFacetAssertions.assertLabelContains(TimeframeFacetSelectors, label);
       CommonAssertions.assertContainsComponentError(
         TimeframeFacetSelectors,
         false
       );
+      CommonFacetAssertions.assertLabelContains(TimeframeFacetSelectors, label);
       CommonFacetAssertions.assertNumberOfIdleLinkValues(
         TimeframeFacetSelectors,
-        customNumberOfValues
+        1
       );
-      TimeframeFacetAssertions.assertFacetValueContainsText(
-        selectionIndex,
-        `Next 1 ${periodFrames[selectionIndex].unit}`
+      CommonFacetAssertions.assertFirstValueContains(
+        TimeframeFacetSelectors,
+        `Next ${periodFrames[0].amount} years`
       );
-    });
-
-    describe('when selecting a value', () => {
-      function setupSelectLinkValue() {
-        setupTimeframeFacetCustomAmount();
-        selectIdleLinkValueAt(TimeframeFacetSelectors, selectionIndex);
-        cy.wait(TestFixture.interceptAliases.Search);
-      }
-      describe('verify rendering', () => {
-        before(setupSelectLinkValue);
-        CommonFacetAssertions.assertDisplayFacet(TimeframeFacetSelectors, true);
-        // CommonAssertions.assertAccessibility(TimeframeFacetSelectors);
-        CommonFacetAssertions.assertDisplayClearButton(
-          TimeframeFacetSelectors,
-          true
-        );
-        CommonFacetAssertions.assertNumberOfSelectedLinkValues(
-          TimeframeFacetSelectors,
-          1
-        );
-        CommonFacetAssertions.assertNumberOfIdleLinkValues(
-          TimeframeFacetSelectors,
-          customNumberOfValues - 1
-        );
-      });
-
-      describe('verify analytic', () => {
-        before(setupSelectLinkValue);
-        TimeframeFacetAssertions.assertLogTimeframeFacetSelect(
-          field,
-          periodFrames[selectionIndex].unit,
-          period
-        );
-      });
+      CommonFacetAssertions.assertNumberOfSelectedLinkValues(
+        TimeframeFacetSelectors,
+        0
+      );
+      CommonFacetAssertions.assertDisplayClearButton(
+        TimeframeFacetSelectors,
+        false
+      );
     });
   });
 
@@ -525,7 +414,7 @@ describe('Timeframe Facet V1 Test Suites', () => {
     before(() => {
       new TestFixture()
         .with(addTimeframeFacet({label, field}, unitFrames))
-        .withHash(`df[${field}]=past-1-week..now`)
+        .withHash(`df[${field}]=past-1-month..now`)
         .init();
     });
 
@@ -533,10 +422,6 @@ describe('Timeframe Facet V1 Test Suites', () => {
     CommonFacetAssertions.assertNumberOfSelectedLinkValues(
       TimeframeFacetSelectors,
       1
-    );
-    CommonFacetAssertions.assertNumberOfIdleLinkValues(
-      TimeframeFacetSelectors,
-      defaultNumberOfValues - 1
     );
     TimeframeFacetAssertions.assertFacetValueContainsText(
       2,