+++ packages/atomic/src/components/facets/atomic-timeframe-facet/atomic-timeframe-facet.tsx
@@ -104,6 +104,16 @@ export class AtomicTimeframeFacet
    * Specifies if the facet is collapsed.
    */
   @Prop({reflect: true, mutable: true}) public isCollapsed = false;
+  /**
+   * Whether to exclude the parents of folded results when estimating the result count for each facet value.
+   */
+  @Prop() public filterFacetCount = true;
+  /**
+   * The maximum number of results to scan in the index to ensure that the facet lists all potential facet values.
+   * Note: A high injectionDepth may negatively impact the facet request performance.
+   * Minimum: `0`
+   */
+  @Prop() public injectionDepth = 1000;
 
   public initialize() {
     this.manualTimeframes = this.getManualTimeframes();
@@ -119,6 +129,8 @@ export class AtomicTimeframeFacet
       currentValues: this.currentValues,
       generateAutomaticRanges: false,
       sortCriteria: 'descending',
+      filterFacetCount: this.filterFacetCount,
+      injectionDepth: this.injectionDepth,
     };
     this.facet = buildDateFacet(this.bindings.engine, {options});
     this.facetId = this.facet.state.facetId;
@@ -154,10 +166,15 @@ export class AtomicTimeframeFacet
 
   private get currentValues(): DateRangeRequest[] {
     return this.manualTimeframes.map(({period, amount, unit}) =>
-      buildDateRange({
-        start: {period, unit, amount},
-        end: {period: 'now'},
-      })
+      period === 'past'
+        ? buildDateRange({
+            start: {period, unit, amount},
+            end: {period: 'now'},
+          })
+        : buildDateRange({
+            start: {period: 'now'},
+            end: {period, unit, amount},
+          })
     );
   }
 
@@ -214,7 +231,11 @@ export class AtomicTimeframeFacet
 
   private formatFacetValue(facetValue: DateFacetValue) {
     try {
-      const relativeDate = deserializeRelativeDate(facetValue.start);
+      const startDate = deserializeRelativeDate(facetValue.start);
+      const relativeDate =
+        startDate.period === 'past'
+          ? startDate
+          : deserializeRelativeDate(facetValue.end);
       const timeframe = this.getManualTimeframes().find(
         (timeframe) =>
           timeframe.period === relativeDate.period &&