+++ packages/headless/src/features/facets/facet-search-set/generic/generic-facet-search-actions.ts
@@ -42,6 +42,7 @@ export const executeFacetSearch = createAsyncThunk<
         state as StateNeededForCategoryFacetSearch
       );
     }
+
     const response = await searchAPIClient.facetSearch(req);
     dispatch(logFacetSearch(facetId));