+++ packages/headless/src/features/facets/range-facets/generic/range-facet-controller-actions.test.ts
@@ -16,14 +16,7 @@ describe('range facet controller actions', () => {
 
     expect(engine.actions).toEqual([
       expect.objectContaining({
-        type: 'rangeFacet/executeToggleSelect/pending',
-      }),
-      expect.objectContaining({
-        type: 'facetOptions/update',
-        payload: {freezeFacetOrder: true},
-      }),
-      expect.objectContaining({
-        type: 'search/executeSearch/pending',
+        type: 'rangeFacet/executeToggleSelect',
       }),
     ]);
   });