+++ packages/headless/src/state/state-sections.ts
@@ -29,6 +29,8 @@ import {QuestionAnsweringState} from '../features/question-answering/question-an
 import {StandaloneSearchBoxSetState} from '../features/standalone-search-box-set/standalone-search-box-set-state';
 import {RecentResultsState} from '../features/recent-results/recent-results-state';
 import {RecentQueriesState} from '../features/recent-queries/recent-queries-state';
+import {ProductListingState} from '../features/product-listing/product-listing-state';
+import {SortState} from '../features/sort/sort-state';
 
 export interface QuerySection {
   /**
@@ -241,6 +243,20 @@ export interface ProductRecommendationsSection {
    * The information related to the product recommendations endpoint.
    */
   productRecommendations: ProductRecommendationsState;
+}
+
+export interface ProductListingSection {
+  /**
+   * The information related to the product listing endpoint.
+   */
+  productListing: ProductListingState;
+}
+
+export interface StructuredSortSection {
+  /**
+   * The information related to sort when using a structured sort format.
+   */
+  sort: SortState;
 }
 
 export interface TriggerSection {